Test-1/requirements.txt
alembic
beautifulsoup4
blinker
bs4
certifi
chardet
click
emoji
et-xmlfile
Flask
Flask-Migrate
Flask-MySQL
flask-serialize
Flask-SQLAlchemy
fpdf
gunicorn
idna
itsdangerous
jdcal
Jinja2
Mako
MarkupSafe
mysql-connector-python
mysqlclient
openpyxl
pandas
permissive-dict
Pillow
protobuf
PyMySQL
python-dateutil
python-editor
pytz
requests
six
soupsieve
SQLAlchemy
urllib3
waitress
Werkzeug
xlrd
zipp